Bring back 'package' command
The package command is the same as the `full` command with
`--package-dir` argument, except that it has a nicer UI. It allows
running the command simply as `explode-js package [DIR]` which will try to
find the `package.json` in the optinal DIR path. By default it will
search the current directory.